requirements.txt
numpy~=1.26.2
matplotlib~=3.8.2
cupy-cuda11x